pyproject.toml
[project]
name = "DISCORD-BOT"
version = "0.1.0"
description = "SUSC Discord Bot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv==1.0.0",
    "py-cord==2.6.1",
    "ruff>=0.11.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
]
